mapping and storing of all known entities

+++ src/Invoices.Core/Data/ICurrentUserIdAccessor.cs
@@ -2,6 +2,6 @@ namespace Invoices.Core.Data
 {
     public interface ICurrentUserIdAccessor
     {
-        int UserId { get; set; }
+        long UserId { get; set; }
     }
 }
